Update the reporters for metrics.rs to accept AnyEntrySink (#58)
✍️ *Description of changes:* Update `MetricReporterBuilder` with a few
changes:
1. I removed the ability to customize the BackgroundQueueBuilder — this
API was unused and is now considered an implementation detail. If you
want to customize the queue behavior, you can instead use the
`metrics_sink` API.
2. I added a `metrics_sink` API and a `metrics_sink_async_shutdown` API
— the metrics_sink API works with existing sinks. The `async_shutdown`
API works with upcoming sinks that are writing to networked destinations
instead of disk. Since we're already in a Tokio task, there is no reason
to have many levels of `spawn_blocking` involved.

This is a breaking change, but these APIs were previously unused.
🔏 *By submitting this pull request*

- [x] I confirm that I've made a best effort attempt to update all
relevant documentation.
- [x] I confirm that my contribution is made under the terms of the
Apache 2.0 license.