make metrics-rs-bridge be metrics.rs version independent  (#86)
📬 *Issue #, if available:*

✍️ *Description of changes:*

The metrics.rs bridge used to be tied to a specific metrics.rs version,
which would mean that metrique-writer major versions would need to be in
correspondence with metrics.rs major versions, which would be annoying.
This PR uses the approach of users manually inputting a metrics.rs
element to remove the major version discrepancy.

In addition, this PR makes FlushWait be Send + Sync, to prevent
annoyances from usage in Tokio. This is a breaking change but is the
right thing to do, since non Sync/Send futures have usability issues in
Tokio, plus the current implementations are Send + Sync.

🔏 *By submitting this pull request*

- [x] I confirm that I've made a best effort attempt to update all
relevant documentation.
- [x] I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------

Co-authored-by: Ariel Ben-Yehuda <arielby@amazon.com>