feat(test-util): derive Clone and Debug for TestEntrySink (#92)
📬 *Issue #, if available:*

n/a

✍️ *Description of changes:*

The two inner fields for `TestEntrySink` are both wrapping `Arcs` and
have `Debug` implemented. It's a minor paper cut to not be able to
easily clone `TestEntrySink`. This came up while I was working on a test
harness that needed to clone and send a particular queue inside a call
that might happen multiple times.

🔏 *By submitting this pull request*

- [x] I confirm that I've made a best effort attempt to update all
relevant documentation.
- [x] I confirm that my contribution is made under the terms of the
Apache 2.0 license.