minimal tokio version, document and check MSRV policy (#126)
📬 *Issue #, if available:*

Fixes #125 

✍️ *Description of changes:*

Declare support for Tokio 1.38. Also do MSRV policy correctly.

The MSRV policy should not matter now, but we might want to increase the
MSRV some time in the future and it's good to be prepared.

Changed the CI to use the exact Rust version in rustup rather than the
short one, since the UI test matches the exact version as well and I'll
prefer no silent test breakage.

🔏 *By submitting this pull request*

- [x] I confirm that I've made a best effort attempt to update all
relevant documentation.
- [x] I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: Ariel Ben-Yehuda <arielby@amazon.com>